Move MOI, SCIP and HiGHS out of the main package into the extensions.
