+++ Project.toml
@@ -9,12 +9,26 @@ DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 FrankWolfe = "f55ce6ea-fdc5-4628-88c5-0087fe54bd30"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
-MathOptSetDistances = "3b969827-a86c-476c-9527-bb6f1a8fbad5"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-SCIP = "82193955-e24f-5292-bf16-6f2c5261a85f"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
+MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
+MathOptSetDistances = "3b969827-a86c-476c-9527-bb6f1a8fbad5"
+SCIP = "82193955-e24f-5292-bf16-6f2c5261a85f"
+
+[weakdeps]
+MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
+MathOptSetDistances = "3b969827-a86c-476c-9527-bb6f1a8fbad5"
+SCIP = "82193955-e24f-5292-bf16-6f2c5261a85f"
+HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
+
+[extensions]
+# name of extension to the left
+# extension dependencies required to load the extension to the right
+# use a list for multiple extension dependencies
+BosciaMathOptInterfaceExt = ["MathOptInterface", "MathOptSetDistances"]
+BosciaSCIPExt = ["MathOptInterface", "MathOptSetDistances", "SCIP"]
+BosciaHiGHSExt = ["MathOptInterface", "MathOptSetDistances", "HiGHS"]
 
 [compat]
 Bonobo = "0.1.3"
@@ -33,4 +47,4 @@ HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "HiGHS", "Distributions"]
+test = ["Test", "HiGHS", "Distributions", "SCIP", "MathOptInterface"]
