+++ ext/BosciaHiGHSExt.jl
@@ -1,0 +1,11 @@
+module BosciaHiGHSExt
+
+using Boscia
+using HiGHS
+using MathOptInterface
+const MOI = MathOptInterface
+const MOIU = MOI.Utilities
+
+import MathOptSetDistances as MOD
+    
+end # module
