+++ src/Boscia.jl
@@ -4,15 +4,9 @@ using FrankWolfe
 import FrankWolfe: compute_extreme_point
 export compute_extreme_point
 using Random
-using SCIP
-import MathOptInterface
 import Bonobo
 using Printf
 using Dates
-const MOI = MathOptInterface
-const MOIU = MOI.Utilities
-
-import MathOptSetDistances as MOD
 
 include("integer_bounds.jl")
 include("blmo_interface.jl")
@@ -27,7 +21,13 @@ include("heuristics.jl")
 include("strong_branching.jl")
 include("utilities.jl")
 include("interface.jl")
-include("MOI_bounded_oracle.jl")
 include("cube_blmo.jl")
+
+# For extensions
+if !isdefined(Base, :get_extension)
+    include("../ext/BosciaMathOptInterfaceExt.jl")
+    include("../ext/BosciaSCIPExt.jl")
+    include("../ext/BosciaHiGHSExt.jl")
+  end
 
 end # module
